Implement Docker registry push in CircleCI configuration
This commit updates the .circleci/config.yml and build.sh files to enable Docker image pushing to the Docker registry.

An important change includes implementing a loop that pushes Docker images to Docker Hub in the CircleCI configuration. The build.sh script now writes image tags to a file which is then used in CircleCI for image pushing.